# Do some more optimisation of the CSV processing